usermanagement/requirements.txt
Flask
flask_cors
requests
Boto3
python-jose